Python_Backend/Py_Backend/Dev_bot/Dev_bot/RAG-chatbot/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1
langchain==0.1.9
numpy
PyPDF2==3.0.1
python-docx==1.1.0
pymilvus
pytesseract
openai
fitz
sentence-transformers
beautifulsoup4
pdfplumber
frontend
tools
PyMuPDF
json5